// src/app/core/api/models/create-room-input-dto.ts
/* tslint:disable */
/* eslint-disable */
export interface CreateRoomInputDto {
  description: string;
  name: string;
  roomType: 'Living Room' | 'Kitchen' | 'Bedroom' | 'Bathroom' | 'Dining Room' | 'Home Office' | 'Garage' | 'Basement' | 'Attic' | 'Guest Room' | 'Hallway' | 'Laundry Room' | 'Pantry' | 'Play Room' | 'Storage Room' | 'Gym' | 'Home Theater' | 'Garden' | 'Balcony' | 'Other';
}